Removed double use of code to inherit a new class to both childs
