+++ JHotDraw/src/main/java/org/jhotdraw/draw/CreationTool.java
@@ -56,7 +56,7 @@ import org.jhotdraw.util.*;
  * <br>2.0 2006-01-14 Changed to support double precision coordinates.
  * <br>1.0 2003-12-01 Derived from JHotDraw 5.4b1.
  */
-public class CreationTool extends AbstractTool {
+public class CreationTool extends TextGenericTool {
 
     /**
      * Attributes to be applied to the created ConnectionFigure.
