+++ JHotDraw/src/main/java/org/jhotdraw/draw/TextGenericTool.java
@@ -1,0 +1,93 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package org.jhotdraw.draw;
+
+import java.awt.event.MouseEvent;
+import java.util.Map;
+import javax.swing.undo.AbstractUndoableEdit;
+import javax.swing.undo.UndoableEdit;
+import org.jhotdraw.util.ResourceBundleUtil;
+
+/**
+ *
+ * @author Zed V2
+ */
+public class TextGenericTool extends AbstractTool {
+   
+    private FloatingTextField textField;
+    private TextHolderFigure typingTarget;
+    private CreationTool creationTool;
+
+    
+    
+    
+    protected void editingCode() {            
+        if (creationTool.createdFigure != null) {
+             getDrawing().remove((Figure)creationTool.getAddedFigure());
+       } else {
+           typingTarget.setText("");
+           typingTarget.changed();
+      }
+                
+    }
+    
+    
+protected void endEdit() {
+        if (typingTarget != null) {
+            typingTarget.willChange();
+
+            final TextHolderFigure editedFigure = typingTarget;
+            final String oldText = typingTarget.getText();
+            final String newText = textField.getText();
+
+            if (newText.length() > 0) {
+                typingTarget.setText(newText);
+            } else { editingCode();
+
+            }
+            UndoableEdit edit = new AbstractUndoableEdit() {
+
+                @Override
+                public String getPresentationName() {
+                    ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.draw.Labels");
+                    return labels.getString("attribute.text.text");
+                }
+
+                @Override
+                public void undo() {
+                    super.undo();
+                    editedFigure.willChange();
+                    editedFigure.setText(oldText);
+                    editedFigure.changed();
+                }
+
+                @Override
+                public void redo() {
+                    super.redo();
+                    editedFigure.willChange();
+                    editedFigure.setText(newText);
+                    editedFigure.changed();
+                }
+            };
+            getDrawing().fireUndoableEditHappened(edit);
+
+            typingTarget.changed();
+            typingTarget = null;
+            
+            textField.endOverlay();
+        }
+    }    
+    @Override
+    public void deactivate(DrawingEditor editor) {
+        super.deactivate(editor); }
+    
+    
+    @Override
+    public void mouseDragged(MouseEvent e) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+    
+}
